Small getDashboard change for code consistency
And linting errors
