+++ packages/portal/backend/src/controllers/AdminController.ts
@@ -16,6 +16,7 @@ import Util from "../../../../common/Util";
 import {Factory} from "../Factory";
 import {AuditLabel, Course, Deliverable, Grade, Person, PersonKind, Repository, Result, Team} from "../Types";
 
+import {ClusteredResult} from "../../../../common/types/ContainerTypes";
 import {DatabaseController} from "./DatabaseController";
 import {DeliverablesController} from "./DeliverablesController";
 import {GitHubActions} from "./GitHubActions";
@@ -260,6 +261,8 @@ export class AdminController {
                 let testSkip: string[] = [];
                 let testError: string[] = [];
 
+                let cluster: ClusteredResult;
+
                 if (typeof result.output !== 'undefined' && typeof result.output.report !== 'undefined') {
                     const report = result.output.report;
                     if (typeof report.scoreOverall !== 'undefined') {
@@ -284,6 +287,9 @@ export class AdminController {
                     if (typeof report.errorNames !== 'undefined') {
                         testError = report.errorNames;
                     }
+                    if (typeof report.cluster !== 'undefined') {
+                        cluster = report.cluster;
+                    }
                 }
 
                 const resultTrans: AutoTestDashboardTransport = {
@@ -303,7 +309,7 @@ export class AdminController {
                     testError: testError,
                     testSkip:  testSkip,
 
-                    cluster: result.output.report.cluster
+                    cluster: cluster
                 };
 
                 // just return the first result for a repo, unless they are specified
